+++ src/utils/file_utils.c
@@ -487,6 +487,18 @@ extern int _getdrive(void);
 extern int _chdrive(int drive);
 #endif
 
+void file_copy_file(const char *src_path, const char *dst_path, bool overwrite)
+{
+	assert(src_path);
+	assert(dst_path);
+#if (_MSC_VER)
+	CopyFileW(win_utf8to16(src_path), win_utf8to16(dst_path), !overwrite);
+#else
+	const char *cmd = "cp %s %s %s";
+	execute_cmd(str_printf(cmd, !overwrite ? "--update=none" : "--update=all", src_path, dst_path), true, NULL);
+#endif
+}
+
 bool file_delete_file(const char *path)
 {
 	assert(path);